backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
deprecation==2.1.0
Django==5.1.4
django-cors-headers==4.6.0
django-rest-knox==5.0.2
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
frozenlist==1.5.0
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
load-dotenv==0.1.0
multidict==6.1.0
packaging==24.2
postgrest==0.18.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.0.6
replicate==1.0.4
requests==2.32.3
resend==2.4.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
yarl==1.18.3
